Added session ticket encryption, using a key that is either passed by the application, or randomly created for one server instance if not.
Also, fixed a number of 64 bits compile warnings.

+++ picoquic/frames.c
@@ -1977,7 +1977,7 @@ void picoquic_process_possible_ack_of_ack_frame(picoquic_cnx_t * cnx, picoquic_p
     picoquic_cnx_t * pcnx = cnx;
 
 	/* Get the packet type */
-    ret = picoquic_parse_packet_header(cnx->quic, p->bytes, p->length, NULL,
+    ret = picoquic_parse_packet_header(cnx->quic, p->bytes, (uint32_t)p->length, NULL,
         ((cnx->quic->flags&picoquic_context_server) == 0) ? 1 : 0, &ph, &pcnx);
 
     if (ret == 0 && ph.ptype == picoquic_packet_0rtt_protected)
